requirements.txt
pillow
requests
selenium
webdriver-manager